# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NSE Finance API",
    "description": "Python FastAPI replacement for NestJS 9 backend. Keeps MySQL finance_app (localhost:3306) and frontend compatibility GET /stocks?duration=... and GET /stock-exchange/Nse. New domain: portfolios / positions / trades / price_snapshots.",
    "version": "0.1.0"
  },
  "paths": {
    "/portfolios": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Portfolios",
        "operationId": "list_portfolios_portfolios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioRead"
                  },
                  "type": "array",
                  "title": "Response List Portfolios Portfolios Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Create Portfolio",
        "operationId": "create_portfolio_portfolios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portfolios/{portfolio_id}": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Get Portfolio",
        "operationId": "get_portfolio_portfolios__portfolio_id__get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portfolios/{portfolio_id}/baseline": {
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Set Baseline",
        "operationId": "set_baseline_portfolios__portfolio_id__baseline_post",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "baseline_value",
            "in": "query",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "exclusiveMinimum": 0.0
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                }
              ],
              "description": "New baseline value, e.g. 100",
              "title": "Baseline Value"
            },
            "description": "New baseline value, e.g. 100"
          },
          {
            "name": "baseline_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "exclusiveMinimum": 0.0
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional baseline price for positions",
              "title": "Baseline Price"
            },
            "description": "Optional baseline price for positions"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portfolios/{portfolio_id}/performance": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Portfolio Performance",
        "operationId": "portfolio_performance_portfolios__portfolio_id__performance_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "normalized",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, pnl_normalized = pnl / baseline_value * 100",
              "default": true,
              "title": "Normalized"
            },
            "description": "If true, pnl_normalized = pnl / baseline_value * 100"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PerformancePoint"
                  },
                  "title": "Response Portfolio Performance Portfolios  Portfolio Id  Performance Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/stocks/{stock_id}": {
      "patch": {
        "tags": [
          "stocks"
        ],
        "summary": "Update stock",
        "operationId": "update_stock_stocks__stock_id__patch",
        "parameters": [
          {
            "name": "stock_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stock Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Stock Stocks  Stock Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "stocks"
        ],
        "summary": "Delete stock",
        "operationId": "delete_stock_stocks__stock_id__delete",
        "parameters": [
          {
            "name": "stock_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stock Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Stock Stocks  Stock Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/stock-exchange/{symbol}": {
      "get": {
        "tags": [
          "stock-exchange"
        ],
        "summary": "Get Last Traded Price",
        "description": "Mirrors StockExchangeController.getLastTradedPriceOfStock(:id).",
        "operationId": "get_last_traded_price_stock_exchange__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Last Traded Price Stock Exchange  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PerformancePoint": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Qty"
          },
          "avg_buy_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Avg Buy Price"
          },
          "current_price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Price"
          },
          "pnl": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pnl"
          },
          "pnl_normalized": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pnl Normalized"
          },
          "baseline_price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Price"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "qty",
          "avg_buy_price",
          "current_price",
          "pnl",
          "pnl_normalized",
          "baseline_price"
        ],
        "title": "PerformancePoint"
      },
      "PortfolioCreate": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "baseline_value": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Baseline Value",
            "default": "100"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name"
        ],
        "title": "PortfolioCreate"
      },
      "PortfolioRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "baseline_value": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Baseline Value"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "baseline_value",
          "created_at"
        ],
        "title": "PortfolioRead"
      },
...
      "StockUpdate": {
        "properties": {
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransactionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "stockName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stockname"
          }
        },
        "type": "object",
        "title": "StockUpdate"
      },
...
      "TransactionType": {
        "type": "string",
        "enum": [
          "buy",
          "sell"
        ],
        "title": "TransactionType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}